refactor, particle config, fixes
