+++ canvas-controller/hello-world-path.js
