+++ index.html
@@ -23,7 +23,7 @@
       }
     </script>
     <script type="module" src="index.js"></script>
-    <script type="module" src="canvas-bg-controller.js"></script>
+    <script type="module" src="/canvas-controller/main.js"></script>
   </head>
   <body>
     <div id="wrapper">
